requirements.txt
# Google Ads API
google-ads==30.0.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

# Data processing
pandas==2.2.1
numpy==1.26.4

# Utils
python-dotenv==1.0.1
pyyaml==6.0.1

# Export
openpyxl==3.1.2
gspread==6.1.2

# Web dashboard
fastapi==0.115.6
uvicorn[standard]==0.32.1
jinja2==3.1.4
